Add modifiers to strip fields from CRDs during conversion
This allows a consumer of the CRD pkg to modify the
CRD before conversion to avoid any errors that may
be a result of fields they do not care about. It
specifically implements support for stripping
labels, annotations, and conversion, but a user
may pass their own modifiers as well.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>