Do not store CRDs more than once
Currently if a repo has the same CRD (specified by GVK) in a repo
multiple times it will be stored multiple times, distinguished by
filepath. This changes CRDs to only be stored once, with the last parsed
representation serving as the source of truth (if representations vary).
This fixes the issue of having CRDs represented multiple times for a
repo, but does not guarantee that the correct schema is being used. This
should be addressed in a future PR by introducing a config file.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>