Return raw CRDs as with v1 schema
In previous iterations of doc, CRDs were returned with the same bytes as
were present in the repo they were indexed from. However, CRDs are now
stored in the internal CRD schema format, which most tools are not
prepared to parse. The benefit of this change was that we have ability
to do exactly what this PR does: choose to to serve v1 flavors of all
CRDs. In the future it may be useful to have the option to also serve
v1beta1 CRDs, but this should unblock most consumers for the time-being.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>